auth
Created functional 'Login' and 'Logout' links in navbar.
Links are scoped so that you can only see 'Sign Up' or 'Login'
when user is not logged in. Only 'Logout' link shows when
user is logged in. Further work needed to implement Twitter auth.
'Login' form also was created. Need to style navbar.